services/analytics/requirements.txt
asyncio
google-genai
grpcio==1.74.0
grpcio-tools==1.74.0
grpcio-reflection
logging
matplotlib
numpy
pandas
plotly
scikit-learn
shapely
supabase